All id`s are now Int

+++ src/main/scala-2.12/com/mmop/db/models/Account.scala
@@ -7,7 +7,7 @@ import scalikejdbc._
 
 class Account(val fields: Map[String, Any]) extends AbstractModel {
 
-  def id = getField[Long]("id")
+  def id = getField[Int]("id")
   def userId = getField[Int]("userId")
   def amount : BigDecimal = getFieldDecimal("amount")
   def creditLimit : BigDecimal = getFieldDecimal("creditLimit")
